+++ src/main/java/com/sparta/northwind/controllers/EmployeeController.java
@@ -1,12 +1,16 @@
 package com.sparta.northwind.controllers;
 
+import com.sparta.northwind.EmployeeDto;
+import com.sparta.northwind.OrderDto;
 import com.sparta.northwind.entities.Employee;
+import com.sparta.northwind.entities.Order;
 import com.sparta.northwind.repositories.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,21 +20,68 @@ public class EmployeeController {
 	private EmployeeRepository repository;
 
 	@GetMapping("/employees")
-	public List<Employee> getEmployees(){
-		return repository.findAll();
+	public List<EmployeeDto> getEmployees() {
+		List<Employee> employees = repository.findAll();
+		List<EmployeeDto> orderDtos = new ArrayList<>();
+		for (Employee employee : employees) {
+			orderDtos.add(new EmployeeDto(employee.getId(),
+					employee.getLastName(),
+					employee.getFirstName(),
+					employee.getTitle(),
+					employee.getTitleOfCourtesy(),
+					employee.getBirthDate(),
+					employee.getHireDate(),
+					employee.getAddress(),
+					employee.getCity(),
+					employee.getRegion(),
+					employee.getPostalCode(),
+					employee.getCountry(),
+					employee.getHomePhone(),
+					employee.getExtension(),
+					employee.getPhoto(),
+					employee.getNotes(),
+					employee.getPhotoPath(),
+					employee.getSalary()));
+
+		}
+		return orderDtos;
 	}
+
 	@GetMapping("/employees/{id}")
-	public ResponseEntity<Employee> getEmployee(@PathVariable int id){
+	public ResponseEntity<EmployeeDto> getEmployee(@PathVariable int id) {
 		Optional<Employee> optionalEmployee = repository.findById(id);
-		return optionalEmployee.map(employee -> new ResponseEntity<>(employee, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+		return optionalEmployee.map(employee -> new ResponseEntity<>(new EmployeeDto(employee.getId(),
+				employee.getLastName(),
+				employee.getFirstName(),
+				employee.getTitle(),
+				employee.getTitleOfCourtesy(),
+				employee.getBirthDate(),
+				employee.getHireDate(),
+				employee.getAddress(),
+				employee.getCity(),
+				employee.getRegion(),
+				employee.getPostalCode(),
+				employee.getCountry(),
+				employee.getHomePhone(),
+				employee.getExtension(),
+				employee.getPhoto(),
+				employee.getNotes(),
+				employee.getPhotoPath(),
+				employee.getSalary()), HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
 	}
+
 	@DeleteMapping("/employees/{id}")
-	public ResponseEntity<Employee> deleteEmployee(@PathVariable int id){
-		repository.deleteById(id);
-		return new ResponseEntity<>(HttpStatus.OK);
+	public ResponseEntity<Employee> deleteEmployee(@PathVariable int id) {
+		if (repository.existsById(id)) {
+			repository.deleteById(id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
+
 	@PostMapping("/employees")
-	public Employee createEmployee(@RequestBody Employee employee){
-		return repository.save(employee);
+	public ResponseEntity<Employee> createEmployee(@RequestBody Employee employee) {
+		return new ResponseEntity<>(repository.save(employee), HttpStatus.CREATED);
 	}
 }
