+++ src/main/java/com/sparta/northwind/controllers/OrderController.java
@@ -1,5 +1,6 @@
 package com.sparta.northwind.controllers;
 
+import com.sparta.northwind.OrderDto;
 import com.sparta.northwind.entities.Order;
 import com.sparta.northwind.repositories.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,8 +8,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
 //TODO Reference to CustomerID, EmployeeID and ShipVia.
 @RestController
 public class OrderController {
@@ -16,21 +19,57 @@ public class OrderController {
 	private OrderRepository repository;
 
 	@GetMapping("/orders")
-	public List<Order> getOrders(){
-		return repository.findAll();
+	public List<OrderDto> getOrders() {
+		List<Order> orderEntities = repository.findAll();
+		List<OrderDto> orderDtos = new ArrayList<>();
+		for (Order order : orderEntities) {
+			orderDtos.add(new OrderDto(order.getId(),
+					order.getOrderDate(),
+					order.getRequiredDate(),
+					order.getShippedDate(),
+					order.getFreight(),
+					order.getShipName(),
+					order.getShipAddress(),
+					order.getShipCity(),
+					order.getShipRegion(),
+					order.getShipPostalCode(),
+					order.getShipCountry()));
+
+		}
+		return orderDtos;
 	}
+
 	@GetMapping("/orders/{id}")
-	public ResponseEntity<Order> getOrder(@PathVariable int id){
+	public ResponseEntity<OrderDto> getOrder(@PathVariable int id) {
 		Optional<Order> optionalOrder = repository.findById(id);
-		return optionalOrder.map(order -> new ResponseEntity<>(order, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+		return optionalOrder.map(order -> new ResponseEntity<>(new OrderDto(order.getId(),
+				order.getOrderDate(),
+				order.getRequiredDate(),
+				order.getShippedDate(),
+				order.getFreight(),
+				order.getShipName(),
+				order.getShipAddress(),
+				order.getShipCity(),
+				order.getShipRegion(),
+				order.getShipPostalCode(),
+				order.getShipCountry()), HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
 	}
+
 	@DeleteMapping("/orders/{id}")
-	public ResponseEntity<Order> deleteOrder(@PathVariable int id){
-		repository.deleteById(id);
-		return new ResponseEntity<>(HttpStatus.OK);
+	public ResponseEntity<Order> deleteOrder(@PathVariable int id) {
+		if (repository.existsById(id)){
+			repository.deleteById(id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		}
+		else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
+
 	@PostMapping("/orders")
-	public Order createOrder(@RequestBody Order order){
-		return repository.save(order);
+	public ResponseEntity<Order> createOrder(@RequestBody Order order) {
+		return new ResponseEntity<>(repository.save(order),HttpStatus.CREATED);
+
+
 	}
 }
